Refine contribution #4065
* Fix tests
* Update Javadocs
* Update year in license headers
* Move observability APIs to a separate package
