+++ pom.xml
@@ -55,7 +55,6 @@
 		<spring-retry.version>1.3.1</spring-retry.version>
 		<spring-integration.version>6.0.0-SNAPSHOT</spring-integration.version>
 		<micrometer.version>2.0.0-SNAPSHOT</micrometer.version>
-		<micrometer-tracing.version>1.0.0-SNAPSHOT</micrometer-tracing.version>
 		<jackson.version>2.13.1</jackson.version>
 
 		<!-- optional production dependencies -->
@@ -80,6 +79,7 @@
 		<junit-jupiter.version>5.8.2</junit-jupiter.version>
 
 		<!-- test dependencies -->
+		<micrometer-tracing.version>1.0.0-SNAPSHOT</micrometer-tracing.version>
 		<junit.version>4.13.2</junit.version>
 		<hamcrest.version>2.2</hamcrest.version>
 		<assertj.version>3.21.0</assertj.version>
