--- spring-batch-core/src/main/java/org/springframework/batch/core/job/BatchJobTagsProvider.java
@@ -1,40 +1,0 @@
-/*
- * Copyright 2006-2009 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.springframework.batch.core.job;
-
-import io.micrometer.core.instrument.observation.Observation;
-
-import org.springframework.batch.core.Job;
-import org.springframework.batch.core.JobExecution;
-import org.springframework.batch.core.JobInterruptedException;
-import org.springframework.batch.core.StartLimitExceededException;
-import org.springframework.batch.core.Step;
-import org.springframework.batch.core.StepExecution;
-import org.springframework.batch.core.repository.JobRestartException;
-
-/**
- * {@link Observation.TagsProvider} for {@link BatchJobContext}.
- * 
- * @author Marcin Grzejszczak
- */
-public interface BatchJobTagsProvider extends Observation.TagsProvider<BatchJobContext> {
-
-	@Override
-	default boolean supportsContext(Observation.Context context) {
-		return context instanceof BatchJobContext;
-	}
-}
