+++ spring-batch-core/src/main/java/org/springframework/batch/core/observability/DefaultBatchStepTagsProvider.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2011-2018 the original author or authors.
+ * Copyright 2022 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.batch.core.step;
+package org.springframework.batch.core.observability;
 
 import io.micrometer.core.instrument.Tags;
 
@@ -23,8 +23,10 @@ import org.springframework.batch.core.StepExecution;
  * Default {@link BatchStepTagsProvider} implementation.
  *
  * @author Marcin Grzejszczak
+ * @since 5.0
  */
 public class DefaultBatchStepTagsProvider implements BatchStepTagsProvider {
+
 	@Override
 	public Tags getLowCardinalityTags(BatchStepContext context) {
 		StepExecution execution = context.getStepExecution();
