--- spring-batch-core/src/main/java/org/springframework/batch/core/step/BatchStepTagsProvider.java
@@ -1,32 +1,0 @@
-/*
- * Copyright 2006-2009 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.springframework.batch.core.step;
-
-import io.micrometer.core.instrument.observation.Observation;
-
-/**
- * {@link Observation.TagsProvider} for {@link BatchStepContext}.
- *
- * @author Marcin Grzejszczak
- */
-public interface BatchStepTagsProvider extends Observation.TagsProvider<BatchStepContext> {
-
-	@Override
-	default boolean supportsContext(Observation.Context context) {
-		return context instanceof BatchStepContext;
-	}
-}
